refactor: split monolithic MCP tests into domain-specific files
