+++ nextmv/tests/cli/mcp/conftest.py
@@ -1,0 +1,10 @@
+"""Shared fixtures and helpers for MCP tests."""
+
+import re
+
+_ANSI_RE = re.compile(r"\x1b\[[0-9;]*m")
+
+
+def _strip_ansi(text: str) -> str:
+    """Remove ANSI escape sequences from text."""
+    return _ANSI_RE.sub("", text)
